% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosageCurveWrap.R
\name{dosage_curves_2T}
\alias{dosage_curves_2T}
\title{Fit curves}
\usage{
dosage_curves_2T(
  s1_phospho_file,
  s2_phospho_file,
  s1_peptide_file,
  s2_peptide_file,
  ks_network_file,
  prob_threshold = 0.5,
  dosages,
  range_threshold = 0.3,
  s1_tag,
  s2_tag,
  s1_rep1_colnames,
  s1_rep2_colnames,
  s1_rep1_omit,
  s1_rep2_omit,
  s2_rep1_colnames,
  s2_rep2_colnames,
  s2_rep1_omit,
  s2_rep2_omit,
  bandwidth = 0.1,
  lead_phospho_BL,
  lead_peptide_BL,
  order_s,
  both_s_BL,
  output_dir
)
}
\arguments{
\item{s1_phospho_file}{file name of the phosphopeptide data under scenario 1}

\item{s2_phospho_file}{file name of the phosphopeptide data under scenario 2}

\item{s1_peptide_file}{file name of the peptide data under first scenario 1}

\item{s2_peptide_file}{file name of the peptide data under second scenario 2}

\item{ks_network_file}{file name of the kinase substrate network}

\item{prob_threshold}{probablity threshold for phosphosites filtering}

\item{dosages}{number of dosages}

\item{range_threshold}{a number set to be the threshold for range of the fitted curve}

\item{s1_tag}{a name tag for scenario 1}

\item{s2_tag}{a name tag for scenario 2}

\item{s1_rep1_colnames}{colnames of scenario 1 replicate 1 to grep relevant columns}

\item{s1_rep2_colnames}{colnames of scenario 1 replicate 2 to grep relevant columns}

\item{s1_rep1_omit}{colnames of scenario 1 replicate 1 to omit relevant columns}

\item{s1_rep2_omit}{colnames of scenario 1 replicate 2 to omit relevant columns}

\item{s2_rep1_colnames}{colnames of scenario 2 replicate 1 to grep relevant columns}

\item{s2_rep2_colnames}{olnames of scenario 2 replicate 2 to grep relevant columns}

\item{s2_rep1_omit}{colnames of scenario 2 replicate 1 to omit relevant columns}

\item{s2_rep2_omit}{colnames of scenario 2 replicate 2 to omit relevant columns}

\item{bandwidth}{bandwidth for Gaussian kernal the larger the smoothier the curve is}

\item{lead_phospho_BL}{a boolean value for filthering only leading proteins or not}

\item{lead_peptide_BL}{a boolean value for filthering only leading proteins or not}

\item{order_s}{which replicate should be arranged at the first place}

\item{both_s_BL}{a boolean value, show pairs with both protein and phospho or not}

\item{output_dir}{output directory}
}
\description{
Fit curves
}
